Add files via upload
+OOP refactored
+Cleanup actions moved to register_activation_hook to ensure no cleanup is done on get_option error
+added filter for capabilites 

Todo:
Controller class for admin
View separation to files and loading by locate template
better handling of variables in classes. 
ACL class should be more strict and simple
