+++ wp-admin-ui/assets/jquery/ui.datepicker.js
